agent/report_analysis/requirements.txt
pytesseract
pdf2image
Pillow
python-multipart
nltk
typing